Create Revising the select query 2
Explanation of the Query
SELECT NAME: This specifies that we only want to retrieve the values from the NAME column.

FROM CITY: This identifies the table we are querying.

WHERE: This clause filters the records based on two conditions:

COUNTRYCODE = 'USA': Filters for cities located in the United States. Note that the value is enclosed in single quotes because it is a string (VARCHAR2).

AND POPULATION > 120000: Further narrows the results to only include cities with a population strictly greater than 120,000.