Create African Cities
Logic Breakdown
SELECT CITY.NAME: Specifies that we only want the names of the cities from the CITY table.

JOIN COUNTRY ON CITY.COUNTRYCODE = COUNTRY.CODE: This joins the two tables by matching the 3-character country code found in both tables. * WHERE COUNTRY.CONTINENT = 'Africa': This acts as a filter to ensure only cities belonging to countries in the 'Africa' continent are returned.